requirements.txt
Pillow
mypy